+++ frontend/src/metabase/selectors/api/index.ts
@@ -1,0 +1,2 @@
+export * from "./databases";
+export * from "./tables";
